==> Cocrystal prediction/1-Smile.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from rdkit import Chem
from rdkit.Chem import Descriptors
import seaborn as sns
from sklearn.model_selection import train_test_split

negative_df = pd.read_csv('data/1-smiles/negative_smiles.csv')

positive_df = pd.read_csv('data/1-smiles/positive_smiles.csv')

merged_df = pd.concat([negative_df, positive_df], ignore_index=True)

merged_df.to_csv('data/1-smiles/merged_smiles.csv', index=False)

merged_df

df = pd.read_csv('data/1-smiles/merged_smiles.csv')
shuffled_df = df.sample(frac=1).reset_index(drop=True)
train_df, test_df = train_test_split(
    shuffled_df,
    test_size=0.2,
    random_state=42,
    stratify=shuffled_df["label"]
)

# 保存划分后的数据
train_df.to_csv('data/1-smiles/train_set.csv', index=False)
test_df.to_csv('data/1-smiles/test_set.csv', index=False)

test_df
